requirements.txt
tqdm
imageio
matplotlib
pytorch==1.9.0
torchvision==0.9.0
opencv-python
attrdict